Cleanup stories, implement tests for breakdown-multisig component
