+++ src/modules/components/proposal/proposalVoting/proposalVotingBreakdownMultisig/proposalVotingBreakdownMultisig.tsx
@@ -1,4 +1,5 @@
-import { formatterUtils, type ITabsContentProps, NumberFormat, Tabs } from '../../../../../core';
+import { NumberFormat, Tabs, formatterUtils, invariant, type ITabsContentProps } from '../../../../../core';
+import { useOdsModulesContext } from '../../../odsModulesProvider';
 import { ProposalVotingTab } from '../proposalVotingDefinitions';
 import { ProposalVotingProgress } from '../proposalVotingProgress';
 
@@ -18,7 +19,14 @@ export interface IProposalVotingBreakdownMultisigProps extends Omit<ITabsContent
 }
 
 export const ProposalVotingBreakdownMultisig: React.FC<IProposalVotingBreakdownMultisigProps> = (props) => {
-    const { approvalsAmount, minApprovals, membersCount } = props;
+    const { approvalsAmount, minApprovals, membersCount, ...otherProps } = props;
+
+    const { copy } = useOdsModulesContext();
+
+    invariant(
+        membersCount > 0,
+        'ProposalVotingBreakdownMultisig: memberCount must be a positive number greater than 0',
+    );
 
     const approvalsAmountPercentage = (approvalsAmount / membersCount) * 100;
     const minApprovalsPercentage = (minApprovals / membersCount) * 100;
@@ -29,13 +37,16 @@ export const ProposalVotingBreakdownMultisig: React.FC<IProposalVotingBreakdownM
     const formattedMembersCount = formatterUtils.formatNumber(membersCount, { format: NumberFormat.GENERIC_SHORT });
 
     return (
-        <Tabs.Content value={ProposalVotingTab.BREAKDOWN} {...props}>
+        <Tabs.Content value={ProposalVotingTab.BREAKDOWN} {...otherProps}>
             <ProposalVotingProgress.Container>
                 <ProposalVotingProgress.Item
-                    name="Minimum Approval"
+                    name={copy.proposalVotingBreakdownMultisig.name}
                     value={approvalsAmountPercentage}
                     indicator={minApprovalsPercentage}
-                    description={{ value: formattedApprovalsAmount, text: `of ${formattedMembersCount} Members` }}
+                    description={{
+                        value: formattedApprovalsAmount,
+                        text: copy.proposalVotingBreakdownMultisig.description(formattedMembersCount),
+                    }}
                     showStatusIcon={true}
                     variant={approvalsAmount >= minApprovals ? 'primary' : 'neutral'}
                 />
